#278 [feat] 강제업데이트 로직 수정
